задача 2

+++ 3-2.py
@@ -1,5 +1,5 @@
-number1 = int(input('Введите первое число '))
-number2 = int(input('Введите второе число '))
-number3 = int(input('Введите третье число '))
+number1 = float(input('Введите первое число '))
+number2 = float(input('Введите второе число '))
+number3 = float(input('Введите третье число '))
 decision = (number1+number2+number3)/3
 print(round(decision, 3))
